Baby Care Bot/requirements.txt
openai
python-telegram-bot==13.4